:triangular_flag_on_post: (AuthController) API 주소 변경
인증 관련된 회원 api는 auth로 묶었습니다.
