+++ src/auth/auth.module.ts
@@ -4,10 +4,17 @@ import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserRepository } from '../user/user.repository';
 import { LocalStrategy } from './strategies/local.strategy';
+import { AuthController } from './auth.controller';
+import { TokenModule } from '../token/token.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([UserRepository]), PassportModule],
+  imports: [
+    TypeOrmModule.forFeature([UserRepository]),
+    PassportModule,
+    TokenModule,
+  ],
   providers: [AuthService, LocalStrategy],
+  controllers: [AuthController],
   exports: [AuthService],
 })
 export class AuthModule {}
